refactor : ManualLotto 클래스 생성
